Move instrumentation & metrics helpers to modules
Splitting the Appsignal module makes it a lot easier to read and change,
as it's no longer 700+ lines of helpers and docs.

These helpers are still included in the Appsignal module, but grouping
them in their own modules makes the Appsignal module and the helper
modules easier to maintain.
