+++ .rubocop_todo.yml
@@ -1,12 +1,12 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2018-06-01 17:12:10 +0200 using RuboCop version 0.50.0.
+# on 2019-03-12 21:08:54 +0100 using RuboCop version 0.50.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
 # versions of RuboCop, may require this file to be generated again.
 
-# Offense count: 44
+# Offense count: 50
 Metrics/AbcSize:
   Max: 56
 
@@ -28,13 +28,13 @@ Metrics/ClassLength:
 Metrics/CyclomaticComplexity:
   Max: 11
 
-# Offense count: 523
+# Offense count: 532
 # Configuration parameters: AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, IgnoredPatterns.
 # URISchemes: http, https
 Metrics/LineLength:
   Max: 152
 
-# Offense count: 94
+# Offense count: 109
 # Configuration parameters: CountComments.
 Metrics/MethodLength:
   Max: 60
@@ -52,7 +52,7 @@ Metrics/PerceivedComplexity:
 # Offense count: 9
 Naming/AccessorMethodName:
   Exclude:
-    - 'lib/appsignal.rb'
+    - 'lib/appsignal/helpers/instrumentation.rb'
     - 'lib/appsignal/transaction.rb'
 
 # Offense count: 4
@@ -71,6 +71,7 @@ Security/YAMLLoad:
 Style/Alias:
   Exclude:
     - 'lib/appsignal.rb'
+    - 'lib/appsignal/helpers/instrumentation.rb'
     - 'lib/appsignal/hooks/active_support_notifications.rb'
     - 'lib/appsignal/hooks/celluloid.rb'
     - 'lib/appsignal/hooks/net_http.rb'
@@ -88,7 +89,7 @@ Style/ClassAndModuleChildren:
   Exclude:
     - 'lib/appsignal/integrations/padrino.rb'
 
-# Offense count: 4
+# Offense count: 6
 Style/ClassVars:
   Exclude:
     - 'lib/appsignal/event_formatter.rb'
